requirements.txt
anthropic>=0.25.0
elevenlabs>=1.0.0
gtts>=2.3.0
moviepy==1.0.3
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth>=2.0.0
schedule>=1.2.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9